Queue &export=download chain at end of the last one
Previously, the DL chains were interweaved, which I got to work,
albeit with difficulty. But it turned out that the way I did this -
queueing one when the other hit its confirmation page - didn't work
when there wasn't a confirmation page. (Technically, it's because
download_child_p doesn't get called on 302). So this just makes them
run sequentially.